refactor: move HTTP hooks into a separate package
