+++ agent/src/main/java/com/appland/appmap/process/hooks/remoterecording/ServletHooks.java
@@ -23,6 +23,14 @@ import com.appland.appmap.transform.annotations.HookCondition;
 import com.appland.appmap.transform.annotations.MethodEvent;
 import com.appland.appmap.util.Logger;
 
+/*
+  * ServletHooks installs the hooks that manage remote recording. If there is at
+  * least one filter installed, the hook for Filter.doFilter will handle the
+  * requests for the remote recording endpoint. If there are no filters
+  * installed, the hook for Servlet.service will handle it instead.
+  *
+  * TODO: move the test framework hooks (JUnit, TestNG) somewhere else.
+  */
 
 public class ServletHooks {
   private static final boolean debug = Properties.DebugHttp;
